// packages/filesystem/src/common/filesystem.d.ts
export declare const FileSystem: unique symbol;
// ...
export declare namespace FileAccess {
    namespace Constants {
        /**
         * Flag indicating that the file is visible to the calling process.
         * This is useful for determining if a file exists, but says nothing about rwx permissions. Default if no mode is specified.
         */
        const F_OK: number;
        /**
         * Flag indicating that the file can be read by the calling process.
         */
        const R_OK: number;
        /**
         * Flag indicating that the file can be written by the calling process.
         */
        const W_OK: number;
        /**
         * Flag indicating that the file can be executed by the calling process.
         * This has no effect on Windows (will behave like `FileAccess.F_OK`).
         */
        const X_OK: number;
    }
}
export interface FileMoveOptions {
    overwrite?: boolean;
}
export interface FileDeleteOptions {
    moveToTrash?: boolean;
}
/**
 * A callback type, called when we try to save a file but realize it has been
 * modified by somebody else since we have opened it.  `originalStat` is the
 * stat at the moment we opened the file, `currentStat` is the stat at the
 * moment we try to save it (after the external modification).  The callback
 * should return true if we still want to save the file, false otherwise.
 */
export declare const FileShouldOverwrite: unique symbol;
export interface FileShouldOverwrite {
    (originalStat: FileStat, currentStat: FileStat): Promise<boolean>;
}
export interface FileSystemClient {
    /**
     * Tests whether the given file can be overwritten
     * in the case if it is out of sync with the given file stat.
     */
    shouldOverwrite: FileShouldOverwrite;
    onDidMove(sourceUri: string, targetUri: string): void;
    onWillMove(sourceUri: string, targetUri: string): void;
}
export declare class DispatchingFileSystemClient implements FileSystemClient {
    readonly clients: Set<FileSystemClient>;
    shouldOverwrite(originalStat: FileStat, currentStat: FileStat): Promise<boolean>;
    onDidMove(sourceUri: string, targetUri: string): void;
    onWillMove(sourceUri: string, targetUri: string): void;
}
/**
 * A file resource with meta information.
 */
export interface FileStat {
    /**
     * The URI of the file.
     */
    uri: string;
    /**
     * The last modification of this file.
     */
    lastModification: number;
    /**
     * `true` if the resource is a directory. Otherwise, `false`.
     */
    isDirectory: boolean;
    /**
     * The children of the file stat.
     * If it is `undefined` and `isDirectory` is `true`, then this file stat is unresolved.
     */
    children?: FileStat[];
    /**
     * The size of the file if known.
     */
    size?: number;
}
export declare namespace FileStat {
    function is(candidate: Object | undefined): candidate is FileStat;
    function equals(one: object | undefined, other: object | undefined): boolean;
}
export declare namespace FileSystemError {
    const FileNotFound: ApplicationError.Constructor<-33000, {
        uri: string;
    }>;
    const FileExists: ApplicationError.Constructor<-33001, {
        uri: string;
    }>;
    const FileIsDirectory: ApplicationError.Constructor<-33002, {
        uri: string;
    }>;
    const FileNotDirectory: ApplicationError.Constructor<-33003, {
        uri: string;
    }>;
    const FileIsOutOfSync: ApplicationError.Constructor<-33004, {
        file: FileStat;
        stat: FileStat;
    }>;
}